Move information generator for AUX import to logic
